moved WorkspaceAdapter
