+++ src/Layout/Navigation/__tests__/RecentList.spec.tsx
@@ -18,7 +18,7 @@ import { RenderResult, render, screen } from '@testing-library/react';
 import { Store } from 'redux';
 
 import NavigationRecentList from '../RecentList';
-import { convertWorkspace, Workspace } from '../../../services/helpers/workspaceAdapter';
+import { convertWorkspace, Workspace } from '../../../services/workspaceAdapter';
 import { FakeStoreBuilder } from '../../../store/__mocks__/storeBuilder';
 import { createHashHistory } from 'history';
 import { createFakeCheWorkspace } from '../../../store/__mocks__/workspace';
