+++ src/pages/WorkspacesList/Rows.tsx
@@ -16,7 +16,7 @@ import { IRow, SortByDirection } from '@patternfly/react-table';
 import WorkspaceIndicator from '../../components/Workspace/Indicator';
 import { formatDate, formatRelativeDate } from '../../services/helpers/date';
 import { buildDetailsPath, toHref, buildIdeLoaderPath } from '../../services/helpers/location';
-import { isWorkspaceV1, Workspace } from '../../services/helpers/workspaceAdapter';
+import { isWorkspaceV1, Workspace } from '../../services/workspaceAdapter';
 import { IDevWorkspaceDevfile } from '@eclipse-che/devworkspace-client';
 
 export interface RowData extends IRow {
