+++ src/store/Workspaces/index.ts
@@ -16,7 +16,7 @@ import { container } from '../../inversify.config';
 import { CheWorkspaceClient } from '../../services/workspace-client/cheWorkspaceClient';
 import { createState } from '../helpers';
 import { IDevWorkspaceDevfile } from '@eclipse-che/devworkspace-client';
-import { convertWorkspace, isWorkspaceV2, isDevfileV2, Workspace } from '../../services/helpers/workspaceAdapter';
+import { convertWorkspace, isWorkspaceV2, isDevfileV2, Workspace } from '../../services/workspaceAdapter';
 import * as CheWorkspacesStore from './cheWorkspaces';
 import * as DevWorkspacesStore from './devWorkspaces';
